.opencode/skills/gemini-drive/requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
google-generativeai>=0.8.4
python-dotenv>=1.0.0
